Startup page ui
1.Startup page ui
2.Format code
