+++ Neuron/Resource/Images.xcassets/commonIcon/peckshield.imageset/Contents.json
@@ -2,17 +2,17 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "launchImage.png",
+      "filename" : "Group@1x.png",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "launchImage@2x.png",
+      "filename" : "Group@2x.png",
       "scale" : "2x"
     },
     {
       "idiom" : "universal",
-      "filename" : "launchImage@3x.png",
+      "filename" : "Group@3x.png",
       "scale" : "3x"
     }
   ],
